Move contract tests to contract folder
