+++ test/contract/AssetHolder/transferAll.test.ts
@@ -1,7 +1,7 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import AssetHolderArtifact from '../../build/contracts/TESTAssetHolder.json';
+import AssetHolderArtifact from '../../../build/contracts/TESTAssetHolder.json';
 import {
   setupContracts,
   newAssetTransferredEvent,
@@ -9,8 +9,8 @@ import {
   allocationToParams,
   sendTransaction,
   replaceAddresses,
-} from '../test-helpers';
-import {createTransferAllTransaction} from '../../src/contract/transaction-creators/asset-holder';
+} from '../../test-helpers';
+import {createTransferAllTransaction} from '../../../src/contract/transaction-creators/asset-holder';
 const AssetHolderInterface = new ethers.utils.Interface(AssetHolderArtifact.abi);
 import {id, bigNumberify} from 'ethers/utils';
