+++ test/contract/ETHAssetHolder/deposit.test.ts
@@ -1,10 +1,10 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ETHAssetHolderArtifact from '../../build/contracts/ETHAssetHolder.json';
-import {setupContracts, newDepositedEvent, sendTransaction} from '../test-helpers';
-import {Channel, getChannelId} from '../../src/contract/channel';
-import {createDepositTransaction} from '../../src/contract/transaction-creators/eth-asset-holder';
+import ETHAssetHolderArtifact from '../../../build/contracts/ETHAssetHolder.json';
+import {setupContracts, newDepositedEvent, sendTransaction} from '../../test-helpers';
+import {Channel, getChannelId} from '../../../src/contract/channel';
+import {createDepositTransaction} from '../../../src/contract/transaction-creators/eth-asset-holder';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
