+++ test/contract/ForceMove/_acceptableWhoSignedWhat.test.ts
@@ -1,8 +1,8 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
-import {setupContracts} from '../test-helpers';
+import ForceMoveArtifact from '../../../build/contracts/TESTForceMove.json';
+import {setupContracts} from '../../test-helpers';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
