+++ test/contract/ForceMove/concludeFromChallenge.test.ts
@@ -1,9 +1,9 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
+import ForceMoveArtifact from '../../../build/contracts/TESTForceMove.json';
 // @ts-ignore
-import countingAppArtifact from '../../build/contracts/CountingApp.json';
+import countingAppArtifact from '../../../build/contracts/CountingApp.json';
 import {keccak256, defaultAbiCoder, bigNumberify} from 'ethers/utils';
 import {
   setupContracts,
@@ -11,13 +11,13 @@ import {
   clearedChallengeHash,
   signStates,
   sendTransaction,
-} from '../test-helpers';
-import {AddressZero} from 'ethers/constants';
-import {Channel, getChannelId} from '../../src/contract/channel';
-import {State, hashState} from '../../src/contract/state';
-import {Outcome, hashOutcome} from '../../src/contract/outcome';
-import {ChannelStorage, hashChannelStorage} from '../../src/contract/channel-storage';
-import {createConcludeFromChallengeTransaction} from '../../src/contract/transaction-creators/force-move';
+} from '../../test-helpers';
+import {AddressZero, HashZero} from 'ethers/constants';
+import {Channel, getChannelId} from '../../../src/contract/channel';
+import {State, hashState} from '../../../src/contract/state';
+import {Outcome, hashOutcome} from '../../../src/contract/outcome';
+import {ChannelStorage, hashChannelStorage} from '../../../src/contract/channel-storage';
+import {createConcludeFromChallengeTransaction} from '../../../src/contract/transaction-creators/force-move';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
