+++ test/contract/ForceMove/forceMove.test.ts
@@ -1,9 +1,9 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
+import ForceMoveArtifact from '../../../build/contracts/TESTForceMove.json';
 // @ts-ignore
-import countingAppArtifact from '../../build/contracts/CountingApp.json';
+import countingAppArtifact from '../../../build/contracts/CountingApp.json';
 import {defaultAbiCoder, hexlify} from 'ethers/utils';
 import {HashZero} from 'ethers/constants';
 import {
@@ -15,12 +15,12 @@ import {
   newChallengeRegisteredEvent,
   sendTransaction,
   signStates,
-} from '../test-helpers';
-import {Channel, getChannelId} from '../../src/contract/channel';
-import {State, getVariablePart, getFixedPart} from '../../src/contract/state';
-import {hashChallengeMessage} from '../../src/contract/challenge';
-import {hashChannelStorage, ChannelStorage} from '../../src/contract/channel-storage';
-import {createForceMoveTransaction} from '../../src/contract/transaction-creators/force-move';
+} from '../../test-helpers';
+import {Channel, getChannelId} from '../../../src/contract/channel';
+import {State, getVariablePart, getFixedPart} from '../../../src/contract/state';
+import {hashChallengeMessage} from '../../../src/contract/challenge';
+import {hashChannelStorage, ChannelStorage} from '../../../src/contract/channel-storage';
+import {createForceMoveTransaction} from '../../../src/contract/transaction-creators/force-move';
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
 );
