+++ test/contract/ForceMove/respond.test.ts
@@ -1,17 +1,17 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
+import ForceMoveArtifact from '../../../build/contracts/TESTForceMove.json';
 // @ts-ignore
-import countingAppArtifact from '../../build/contracts/CountingApp.json';
+import countingAppArtifact from '../../../build/contracts/CountingApp.json';
 import {keccak256, defaultAbiCoder, hexlify, toUtf8Bytes, bigNumberify} from 'ethers/utils';
-import {setupContracts, sign, newChallengeClearedEvent, sendTransaction} from '../test-helpers';
+import {setupContracts, sign, newChallengeClearedEvent, sendTransaction} from '../../test-helpers';
 import {HashZero, AddressZero} from 'ethers/constants';
-import {Outcome, hashOutcome} from '../../src/contract/outcome';
-import {Channel, getChannelId} from '../../src/contract/channel';
-import {State, hashState, getVariablePart, getFixedPart} from '../../src/contract/state';
-import {hashChannelStorage} from '../../src/contract/channel-storage';
-import {createRespondTransaction} from '../../src/contract/transaction-creators/force-move';
+import {Outcome, hashOutcome} from '../../../src/contract/outcome';
+import {Channel, getChannelId} from '../../../src/contract/channel';
+import {State, hashState, getVariablePart, getFixedPart} from '../../../src/contract/state';
+import {hashChannelStorage} from '../../../src/contract/channel-storage';
+import {createRespondTransaction} from '../../../src/contract/transaction-creators/force-move';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
