+++ test/contract/ForceMove/respondWithAlternative.test.ts
@@ -1,9 +1,9 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
+import ForceMoveArtifact from '../../../build/contracts/TESTForceMove.json';
 // @ts-ignore
-import countingAppArtifact from '../../build/contracts/CountingApp.json';
+import countingAppArtifact from '../../../build/contracts/CountingApp.json';
 import {keccak256, defaultAbiCoder, hexlify, bigNumberify} from 'ethers/utils';
 import {
   setupContracts,
@@ -11,13 +11,13 @@ import {
   newChallengeClearedEvent,
   signStates,
   sendTransaction,
-} from '../test-helpers';
+} from '../../test-helpers';
 import {AddressZero} from 'ethers/constants';
-import {Outcome} from '../../src/contract/outcome';
-import {Channel, getChannelId} from '../../src/contract/channel';
-import {State} from '../../src/contract/state';
-import {hashChannelStorage} from '../../src/contract/channel-storage';
-import {createRespondWithAlternativeTransaction} from '../../src/contract/transaction-creators/force-move';
+import {Outcome} from '../../../src/contract/outcome';
+import {Channel, getChannelId} from '../../../src/contract/channel';
+import {State} from '../../../src/contract/state';
+import {hashChannelStorage} from '../../../src/contract/channel-storage';
+import {createRespondWithAlternativeTransaction} from '../../../src/contract/transaction-creators/force-move';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
